+++ plom_server/scripts/launch_plom_demo_server.py
@@ -33,7 +33,6 @@ from plom_server.scripts.launch_plom_server import (
     launch_django_dev_server_process,
 )
 
-
 # TODO: not a fan of global variables, and mypy needs this to be defined
 global demo_files
 # so temporarily set it to "."; we'll fix it in main()
@@ -295,7 +294,7 @@ def _build_with_and_without_soln(source_path: Path) -> None:
         print(f"  - skipping build of {no_soln_pdf_filename} b/c it already exists")
     else:
         with open(no_soln_pdf_filename, "wb") as f:
-            (r, stdouterr) = buildLaTeX(no_soln_data, f)
+            r, stdouterr = buildLaTeX(no_soln_data, f)
         if r != 0:
             print(stdouterr)
             raise RuntimeError(
@@ -311,7 +310,7 @@ def _build_with_and_without_soln(source_path: Path) -> None:
         print(f"  - skipping build of {yes_soln_pdf_filename} b/c it already exists")
     else:
         with open(yes_soln_pdf_filename, "wb") as f:
-            (r, stdouterr) = buildLaTeX(yes_soln_data, f)
+            r, stdouterr = buildLaTeX(yes_soln_data, f)
         if r != 0:
             print(stdouterr)
             raise RuntimeError(
@@ -667,20 +666,25 @@ def run_the_auto_id_reader():
 def _ensure_client_available():
     try:
         # tell MyPy to ignore this for testing
-        import plomclient  # type: ignore[import-not-found]
-        from plomclient.client import __version__ as clientversion  # type: ignore
+        from plom.client import __version__ as clientversion  # type: ignore
     except ImportError as err:
-        print("*" * 64)
-        print()
+        # We should be able to delete this "plomclient" try-except-else in say 2027 or so
+        try:
+            from plomclient.client import __version__ as clientversion  # type: ignore
+        except ImportError:
+            pass
+        else:
+            raise RuntimeError(
+                "The randoiding and randomarking utilities depend on plom-client,"
+                f" which is installed but is too old: {clientversion}\n"
+                "Either upgrade plom-client, or stop the demo earlier."
+            ) from None
         raise RuntimeError(
             "The randoiding and randomarking utilities depend on plom-client, "
             f"which is not installed:\n  {err}.\n"
             "Either install plom-client, or stop the demo earlier."
         ) from None
-    print(
-        f"Good we have plom-client installed, version {clientversion},"
-        f" found at {plomclient}"
-    )
+    print(f"Good, we have plom-client installed, version {clientversion}")
 
 
 def run_the_randoider(*, port):
@@ -697,7 +701,7 @@ def run_the_randoider(*, port):
         ("demoMarker1", "demoMarker1"),
     ]
 
-    cmd = f"python3 -m plomclient.client.randoIDer -s {srv} -u {users[0][0]} -w {users[0][1]} --use-predictions"
+    cmd = f"python3 -m plom.client.randoIDer -s {srv} -u {users[0][0]} -w {users[0][1]} --use-predictions"
     print(f"RandoIDing!  calling: {cmd}")
     subprocess.check_call(split(cmd))
 
@@ -722,7 +726,7 @@ def run_the_randomarker(*, port, half_marks=False):
 
     randomarker_processes = []
     for X in users[1:]:
-        cmd = f"python3 -m plomclient.client.randoMarker -s {srv} -u {X[0]} -w {X[1]} --partial {X[2]} --download-rubrics"
+        cmd = f"python3 -m plom.client.randoMarker -s {srv} -u {X[0]} -w {X[1]} --partial {X[2]} --download-rubrics"
         if half_marks:
             cmd += " --allow-half"
         print(f"RandoMarking!  calling: {cmd}")
@@ -745,7 +749,7 @@ def run_the_randomarker(*, port, half_marks=False):
 
     # now a final run to do any remaining tasks
     for X in users[:1]:
-        cmd = f"python3 -m plomclient.client.randoMarker -s {srv} -u {X[0]} -w {X[1]} --partial 100 --download-rubrics"
+        cmd = f"python3 -m plom.client.randoMarker -s {srv} -u {X[0]} -w {X[1]} --partial 100 --download-rubrics"
         if half_marks:
             cmd += " --allow-half"
         print(f"RandoMarking!  calling: {cmd}")
@@ -905,7 +909,9 @@ def main():
     saytime("Finished refreshing the database.")
 
     # build the user-groups and the admin and manager users
-    run_django_manage_command("plom_make_groups_and_first_users")
+    run_django_manage_command(
+        f"plom_make_groups_and_first_users --manager-login manager 1234 --port {args.port}"
+    )
     # build extra-page and scrap-paper PDFs
     run_django_manage_command("plom_build_scrap_extra_pdfs")