+++ CHANGELOG.md
@@ -18,6 +18,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.21.2] - 2026-06-25
+
+### Changed
+* The format of extra sheets has changed.
+
+### Fixed
+* Fix some regressions when creating rubrics via the web UI or from CSV files.
+
+
+
 ## [0.21.1] - 2026-06-04
 
 ### Changed
@@ -1697,7 +1707,8 @@ in most cases.
 This is the first release of Plom, Paperless Open Marking.
 
 
-[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.21.1...main
+[Unreleased]: https://gitlab.com/plom/plom/-/compare/v0.21.2...main
+[0.21.2]: https://gitlab.com/plom/plom/-/compare/v0.21.1...v0.21.2
 [0.21.1]: https://gitlab.com/plom/plom/-/compare/v0.21.0...v0.21.1
 [0.21.0]: https://gitlab.com/plom/plom/-/compare/v0.20.2...v0.21.0
 [0.20.2]: https://gitlab.com/plom/plom/-/compare/v0.20.1...v0.20.2